+++ docs/reference-guides/core-blocks.md
@@ -8,6 +8,17 @@ This page lists the blocks included in the block-library package.
 
 <!-- START TOKEN Autogenerated - DO NOT EDIT -->
 
+## Accordion
+
+A group of Accordion headers and associated expandable content. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/accordion))
+
+-	**Name:** core/accordion
+-	**Experimental:** true
+-	**Category:** design
+-	**Allowed Blocks:** core/accordion-item
+-	**Supports:** align (full, wide), background (backgroundImage, backgroundSize), color (background, gradient, text), interactivity, layout, shadow, spacing (blockGap, margin, padding), ~~html~~
+-	**Attributes:** allowedBlocks, autoclose, iconPosition, showIcon
+
 ## Accordion Header
 
 Accordion header. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/accordion-header))
@@ -19,14 +30,14 @@ Accordion header. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/pa
 -	**Supports:** anchor, border, color (background, gradient, text), interactivity, shadow, spacing (margin, padding), typography (fontSize, textAlign), ~~align~~
 -	**Attributes:** iconPosition, level, levelOptions, openByDefault, showIcon, textAlignment, title
 
-## Accordion
+## Accordion Item
 
 A single accordion that displays a header and expandable content. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/accordion-item))
 
 -	**Name:** core/accordion-item
 -	**Experimental:** true
 -	**Category:** design
--	**Parent:** core/accordions
+-	**Parent:** core/accordion
 -	**Allowed Blocks:** core/accordion-header, core/accordion-panel
 -	**Supports:** align (full, wide), color (background, gradient, text), interactivity, shadow, spacing (blockGap, margin)
 -	**Attributes:** openByDefault
@@ -41,17 +52,6 @@ Accordion Panel ([Source](https://github.com/WordPress/gutenberg/tree/trunk/pack
 -	**Parent:** core/accordion-item
 -	**Supports:** border, color (background, gradient, text), interactivity, shadow, spacing (blockGap, margin, padding), typography (fontSize, lineHeight)
 -	**Attributes:** allowedBlocks, isSelected, openByDefault, templateLock
-
-## Accordions
-
-A group of Accordion headers and associated expandable content. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/accordions))
-
--	**Name:** core/accordions
--	**Experimental:** true
--	**Category:** design
--	**Allowed Blocks:** core/accordion-item
--	**Supports:** align (full, wide), background (backgroundImage, backgroundSize), color (background, gradient, text), interactivity, layout, shadow, spacing (blockGap, margin, padding), ~~html~~
--	**Attributes:** allowedBlocks, autoclose, iconPosition, showIcon
 
 ## Archives
