+++ packages/block-library/src/accordion-item/block.json
@@ -3,12 +3,12 @@
 	"apiVersion": 3,
 	"name": "core/accordion-item",
 	"version": "0.1.0",
-	"title": "Accordion",
+	"title": "Accordion Item",
 	"category": "design",
 	"description": "A single accordion that displays a header and expandable content.",
 	"example": {},
 	"__experimental": true,
-	"parent": [ "core/accordions" ],
+	"parent": [ "core/accordion" ],
 	"allowedBlocks": [ "core/accordion-header", "core/accordion-panel" ],
 	"supports": {
 		"align": [ "wide", "full" ],
