+++ packages/block-library/src/accordion/block.json
@@ -1,8 +1,8 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 3,
-	"name": "core/accordions",
-	"title": "Accordions",
+	"name": "core/accordion",
+	"title": "Accordion",
 	"category": "design",
 	"description": "A group of Accordion headers and associated expandable content.",
 	"example": {},
@@ -65,5 +65,5 @@
 	},
 	"allowedBlocks": [ "core/accordion-item" ],
 	"textdomain": "default",
-	"style": "wp-block-accordions"
+	"style": "wp-block-accordion"
 }
