+++ packages/block-library/src/accordion/index.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Server-side rendering of the `core/accordions` block.
+ * Server-side rendering of the `core/accordion` block.
  *
  * @package WordPress
  * @since 6.6.0
@@ -10,31 +10,31 @@
  *
  * @return string Returns the updated markup.
  */
-function render_block_core_accordions( $attributes, $content ) {
+function render_block_core_accordion( $attributes, $content ) {
 	if ( ! $content ) {
 		return $content;
 	}
 
 	$suffix = wp_scripts_get_suffix();
 	if ( defined( 'IS_GUTENBERG_PLUGIN' ) && IS_GUTENBERG_PLUGIN ) {
-		$module_url = gutenberg_url( '/build-module/block-library/accordions/view.min.js' );
+		$module_url = gutenberg_url( '/build-module/block-library/accordion/view.min.js' );
 	}
 
 	wp_register_script_module(
-		'@wordpress/block-library/accordions',
-		isset( $module_url ) ? $module_url : includes_url( "blocks/accordions/view{$suffix}.js" ),
+		'@wordpress/block-library/accordion',
+		isset( $module_url ) ? $module_url : includes_url( "blocks/accordion/view{$suffix}.js" ),
 		array( '@wordpress/interactivity' ),
 		defined( 'GUTENBERG_VERSION' ) ? GUTENBERG_VERSION : get_bloginfo( 'version' )
 	);
 
-	wp_enqueue_script_module( '@wordpress/block-library/accordions' );
+	wp_enqueue_script_module( '@wordpress/block-library/accordion' );
 
 	$p             = new WP_HTML_Tag_Processor( $content );
 	$autoclose     = $attributes['autoclose'] ? 'true' : 'false';
 	$icon          = $attributes['icon'] ?? 'plus';
 	$icon_position = $attributes['iconPosition'] ?? 'right';
 
-	if ( $p->next_tag( array( 'class_name' => 'wp-block-accordions' ) ) ) {
+	if ( $p->next_tag( array( 'class_name' => 'wp-block-accordion' ) ) ) {
 		$p->set_attribute( 'data-wp-interactive', 'core/accordion' );
 		$p->set_attribute( 'data-wp-context', '{ "autoclose": ' . $autoclose . ', "isOpen": [], "icon": "' . $icon . '", "iconPosition": "' . $icon_position . '" }' );
 
@@ -46,16 +46,16 @@ function render_block_core_accordions( $attributes, $content ) {
 }
 
 /**
- * Registers the `core/accordions` block on server.
+ * Registers the `core/accordion` block on server.
  *
  * @since 6.6.0
  */
-function register_block_core_accordions() {
+function register_block_core_accordion() {
 	register_block_type_from_metadata(
-		__DIR__ . '/accordions',
+		__DIR__ . '/accordion',
 		array(
-			'render_callback' => 'render_block_core_accordions',
+			'render_callback' => 'render_block_core_accordion',
 		)
 	);
 }
-add_action( 'init', 'register_block_core_accordions' );
+add_action( 'init', 'register_block_core_accordion' );
