+++ packages/block-library/src/style.scss
@@ -1,4 +1,4 @@
-@import "./accordions/style.scss";
+@import "./accordion/style.scss";
 @import "./archives/style.scss";
 @import "./avatar/style.scss";
 @import "./audio/style.scss";
